+++ NEWS.md
@@ -1,3 +1,8 @@
+# readtextgrid dev version
+
+* Remove magrittr dependency. (#5)
+
+
 # readtextgrid 0.1.1
 
 * Fixes for CRAN resubmission.