+++ NEWS.md
@@ -1,5 +1,10 @@
 # readtextgrid dev version
 
+Nothing yet.
+
+
+# readtextgrid 0.1.2
+
 * Add `encoding` argument to `read_textgrid()`. (#7, #8, #9, @jonorthwash)
 * Include example UTF-16 textgrid for testing.
 * Remove magrittr dependency. (#5, @JoFrhwld)